Jenkinsfile: temporary disable archive on windows

+++ Jenkinsfile
@@ -734,25 +734,27 @@ class TestAndPackage extends AgentJob {
     }
 
     void archive_artifacts() {
-        def artifacts_dir = "artifacts/${agent.host()}"
-        script.sh """
-           mkdir -p ${artifacts_dir}
-           rm -rf ${artifacts_dir}/*
-           if ls glamoroustoolkit/*.log &> /dev/null; then
-              mkdir ${artifacts_dir}/glamoroustoolkit
-              mv glamoroustoolkit/*.log ${artifacts_dir}/glamoroustoolkit/
-           fi
-           if ls gt-examples/*.log &> /dev/null; then
-              mkdir ${artifacts_dir}/gt-examples
-              mv gt-examples/*.log ${artifacts_dir}/gt-examples/
-           fi
-           if ls gt-releaser/*.log &> /dev/null; then
-              mkdir ${artifacts_dir}/gt-releaser
-              mv gt-releaser/*.log ${artifacts_dir}/gt-releaser/
-           fi
-        """
-        script.archiveArtifacts(artifacts: '${artifacts/**}', allowEmptyArchive: true)
-    }
+        if (platform() != Platform.Windows) {
+            def artifacts_dir = "artifacts/${agent.host()}"
+            script.sh """
+               mkdir -p ${artifacts_dir}
+               rm -rf ${artifacts_dir}/*
+               if ls glamoroustoolkit/*.log &> /dev/null; then
+                  mkdir ${artifacts_dir}/glamoroustoolkit
+                  mv glamoroustoolkit/*.log ${artifacts_dir}/glamoroustoolkit/
+               fi
+               if ls gt-examples/*.log &> /dev/null; then
+                  mkdir ${artifacts_dir}/gt-examples
+                  mv gt-examples/*.log ${artifacts_dir}/gt-examples/
+               fi
+               if ls gt-releaser/*.log &> /dev/null; then
+                  mkdir ${artifacts_dir}/gt-releaser
+                  mv gt-releaser/*.log ${artifacts_dir}/gt-releaser/
+               fi
+            """
+            script.archiveArtifacts(artifacts: '${artifacts/**}', allowEmptyArchive: true)
+            }
+        }
 }
 
 class TestAndPackageWithGemstoneAndPython extends TestAndPackage {
